requirements.txt
requests>=2.31.0
mitreattack-python>=1.0.1
stix2>=3.0.1